trading_marl_ga/requirements.txt
# ==========================================
# Deep Learning & Numerical Computing
# ==========================================
torch>=2.0.0
numpy>=1.24.0

# ==========================================
# Data Processing & Analysis
# ==========================================
pandas>=2.0.0
scipy>=1.10.0

# ==========================================
# Korean Stock Market Data
# ==========================================
FinanceDataReader>=0.9.50
pykrx>=1.0.40
OpenDartReader>=0.2.0

# ==========================================
# Utilities
# ==========================================
python-dotenv>=1.0.0
tqdm>=4.65.0

# ==========================================
# Optional (for Colab)
# ==========================================
# matplotlib>=3.7.0  # 시각화
# seaborn>=0.12.0    # 시각화
# jupyterlab>=4.0.0  # 노트북 환경
